+++ MyPlayground.playground/contents.xcplayground
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='6.0' target-platform='ios' executeOnSourceChanges='false'>
+<playground version='6.0' target-platform='ios' display-mode='raw' executeOnSourceChanges='false'>
     <pages>
         <page name='Tiny Constraints'/>
+        <page name='Sizes'/>
         <page name='Edges'/>
         <page name='Relative Constraints'/>
     </pages>
